caso de uso dos de entrega 2

+++ src/main/java/edu/fiuba/algo3/modelo/Jugador.java
@@ -31,7 +31,7 @@ public class Jugador {
         tiempo.sumarHoras(8);
     }
 
-    public void viajar(float distancia, Tiempo tiempo) {
+    public void viajar(double distancia, Tiempo tiempo) {
         cargo.viajar(distancia, tiempo);
     }
